Rename spotifyservice to spotify
