+++ spotify/spotify.go
@@ -1,4 +1,4 @@
-package spotifyservice
+package spotify
 
 import (
 	"github.com/Henry-Sarabia/scry"
@@ -12,21 +12,21 @@ const (
 	publicPlaylist bool = true
 )
 
-type SpotifyService struct {
+type Service struct {
 	c *spotify.Client
 }
 
-func New(c *spotify.Client) (*SpotifyService, error) {
+func New(c *spotify.Client) (*Service, error) {
 	if c == nil {
 		return nil, errors.New("client pointer is nil")
 	}
-	s := &SpotifyService{c: c}
+	s := &Service{c: c}
 
 	return s, nil
 }
 
-func (sp *SpotifyService) TopArtists() ([]scry.Artist, error) {
-	top, err := sp.c.CurrentUsersTopArtists()
+func (s *Service) TopArtists() ([]scry.Artist, error) {
+	top, err := s.c.CurrentUsersTopArtists()
 	if err != nil {
 		return nil, errors.Wrap(err, "cannot fetch top artists")
 	}
@@ -34,8 +34,8 @@ func (sp *SpotifyService) TopArtists() ([]scry.Artist, error) {
 	return ParseArtists(top.Artists...), nil
 }
 
-func (sp *SpotifyService) TopTracks() ([]scry.Track, error) {
-	top, err := sp.c.CurrentUsersTopTracks()
+func (s *Service) TopTracks() ([]scry.Track, error) {
+	top, err := s.c.CurrentUsersTopTracks()
 	if err != nil {
 		return nil, errors.Wrap(err, "cannot fetch top tracks")
 	}
@@ -43,8 +43,8 @@ func (sp *SpotifyService) TopTracks() ([]scry.Track, error) {
 	return ParseFullTracks(top.Tracks...), nil
 }
 
-func (sp *SpotifyService) RecentTracks() ([]scry.Track, error) {
-	rec, err := sp.c.PlayerRecentlyPlayed()
+func (s *Service) RecentTracks() ([]scry.Track, error) {
+	rec, err := s.c.PlayerRecentlyPlayed()
 	if err != nil {
 		return nil, errors.Wrap(err, "cannot fetch recently played tracks")
 	}
@@ -57,7 +57,7 @@ func (sp *SpotifyService) RecentTracks() ([]scry.Track, error) {
 	return t, nil
 }
 
-func (sp *SpotifyService) Recommendations(seeds []scry.Seed) ([]scry.Track, error) {
+func (s *Service) Recommendations(seeds []scry.Seed) ([]scry.Track, error) {
 	sds, err := ParseSeeds(seeds)
 	if err != nil {
 		return nil, err
@@ -67,7 +67,7 @@ func (sp *SpotifyService) Recommendations(seeds []scry.Seed) ([]scry.Track, erro
 	attr := spotify.NewTrackAttributes().TargetPopularity(popTarget).MaxPopularity(popMax)
 
 	for _, sd := range sds {
-		recs, err := sp.c.GetRecommendations(sd, attr, nil)
+		recs, err := s.c.GetRecommendations(sd, attr, nil)
 		if err != nil {
 			return nil, errors.Wrap(err, "cannot fetch recommendations")
 		}
@@ -79,13 +79,13 @@ func (sp *SpotifyService) Recommendations(seeds []scry.Seed) ([]scry.Track, erro
 	return tracks, nil
 }
 
-func (sp *SpotifyService) Playlist(name string, tracks []scry.Track) (*spotify.FullPlaylist, error) {
-	u, err := sp.c.CurrentUser()
+func (s *Service) Playlist(name string, tracks []scry.Track) (*spotify.FullPlaylist, error) {
+	u, err := s.c.CurrentUser()
 	if err != nil {
 		return nil, errors.Wrap(err, "cannot fetch user")
 	}
 
-	pl, err := sp.c.CreatePlaylistForUser(u.ID, name, "description", publicPlaylist)
+	pl, err := s.c.CreatePlaylistForUser(u.ID, name, "description", publicPlaylist)
 	if err != nil {
 		return nil, errors.Wrap(err, "cannot create playlist")
 	}
@@ -95,7 +95,7 @@ func (sp *SpotifyService) Playlist(name string, tracks []scry.Track) (*spotify.F
 		IDs = append(IDs, spotify.ID(t.ID))
 	}
 
-	_, err = sp.c.AddTracksToPlaylist(pl.ID, IDs...)
+	_, err = s.c.AddTracksToPlaylist(pl.ID, IDs...)
 	if err != nil {
 		return nil, errors.Wrap(err, "cannot add tracks to playlist")
 	}
